+++ Moneo.sln
@@ -15,6 +15,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Moneo.Tests", "Tests\Moneo.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moneo.Bot", "src\Moneo.Bot\Moneo.Bot.csproj", "{3C55E819-54DF-45B2-88F5-EB3A94E29136}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moneo.Contracts", "src\Moneo.Contracts\Moneo.Contracts.csproj", "{22B46A0E-420E-4233-B1FD-7D0773724B20}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moneo.Models", "src\Moneo.Models\Moneo.Models.csproj", "{0200A40A-CF26-4117-B888-28BFBDDFD8D6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moneo.Common", "src\Moneo.Common\Moneo.Common.csproj", "{8891757B-FDD7-4C10-ADCA-00CECF92412A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -37,6 +43,18 @@ Global
 		{3C55E819-54DF-45B2-88F5-EB3A94E29136}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3C55E819-54DF-45B2-88F5-EB3A94E29136}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3C55E819-54DF-45B2-88F5-EB3A94E29136}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22B46A0E-420E-4233-B1FD-7D0773724B20}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{22B46A0E-420E-4233-B1FD-7D0773724B20}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22B46A0E-420E-4233-B1FD-7D0773724B20}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22B46A0E-420E-4233-B1FD-7D0773724B20}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0200A40A-CF26-4117-B888-28BFBDDFD8D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0200A40A-CF26-4117-B888-28BFBDDFD8D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0200A40A-CF26-4117-B888-28BFBDDFD8D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0200A40A-CF26-4117-B888-28BFBDDFD8D6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8891757B-FDD7-4C10-ADCA-00CECF92412A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8891757B-FDD7-4C10-ADCA-00CECF92412A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8891757B-FDD7-4C10-ADCA-00CECF92412A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8891757B-FDD7-4C10-ADCA-00CECF92412A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -46,6 +64,9 @@ Global
 		{F0416414-1F51-43A7-8035-565F1B47B976} = {1C2AAB11-7929-4C83-BA18-70CBBB4FE657}
 		{0725362F-6676-4471-A06A-4982F25D74F2} = {BE78DF45-BDCF-46FB-ACCB-7FC2F9590C03}
 		{3C55E819-54DF-45B2-88F5-EB3A94E29136} = {1C2AAB11-7929-4C83-BA18-70CBBB4FE657}
+		{22B46A0E-420E-4233-B1FD-7D0773724B20} = {1C2AAB11-7929-4C83-BA18-70CBBB4FE657}
+		{0200A40A-CF26-4117-B888-28BFBDDFD8D6} = {1C2AAB11-7929-4C83-BA18-70CBBB4FE657}
+		{8891757B-FDD7-4C10-ADCA-00CECF92412A} = {1C2AAB11-7929-4C83-BA18-70CBBB4FE657}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9A69730A-25AE-421F-A76A-39B03C0A0207}
