+++ Tests/Moneo.Tests/Moneo.Tests.csproj
@@ -1,11 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 
     <IsPackable>false</IsPackable>
+
+    <LangVersion>12</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
@@ -22,7 +24,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\Moneo.Common\Moneo.Common.csproj" />
+    <ProjectReference Include="..\..\src\Moneo.Contracts\Moneo.Contracts.csproj" />
     <ProjectReference Include="..\..\src\moneo.core\Moneo.Core.csproj" />
+    <ProjectReference Include="..\..\src\Moneo.Models\Moneo.Models.csproj" />
   </ItemGroup>
 
 </Project>
