+++ src/Moneo.Bot/BotRequests/BotTextMessageRequest.cs
@@ -1,6 +1,6 @@
 using MediatR;
 
-namespace Moneo.Bot.BotRequests;
+namespace Moneo.Chat.BotRequests;
 
 internal record BotTextMessageRequest(long ConversationId, string Text, bool IsError = false) : IRequest;
 internal record BotGifMessageRequest(long ConversationId, string GifUrl) : IRequest;
