+++ src/Moneo.Bot/BotService.cs
@@ -2,14 +2,16 @@ using MediatR;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Moneo.Bot.BotRequests;
+using Moneo.Chat.BotRequests;
+using Moneo.Models.Chat;
+using Moneo.TaskManagement;
 using Telegram.Bot;
 using Telegram.Bot.Exceptions;
 using Telegram.Bot.Polling;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 
-namespace Moneo.Bot;
+namespace Moneo.Chat;
 
 internal class BotService : BackgroundService, IRequestHandler<BotTextMessageRequest>, IRequestHandler<BotGifMessageRequest>
 {
