+++ src/Moneo.Bot/Moneo.Bot.csproj
@@ -2,9 +2,10 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <LangVersion>12</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
@@ -16,7 +17,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Moneo.Common\Moneo.Common.csproj" />
+    <ProjectReference Include="..\Moneo.Contracts\Moneo.Contracts.csproj" />
     <ProjectReference Include="..\moneo.core\Moneo.Core.csproj" />
+    <ProjectReference Include="..\Moneo.Models\Moneo.Models.csproj" />
   </ItemGroup>
 
 </Project>
