+++ src/Moneo.Contracts/Moneo.Contracts.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <Folder Include="Chat\" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Moneo.Models\Moneo.Models.csproj" />
+    </ItemGroup>
+
+</Project>
