+++ src/Moneo.Models/Chat/MessageDirection.cs
@@ -1,0 +1,7 @@
+namespace Moneo.Models.Chat;
+
+public enum MessageDirection
+{
+    UserToBot,
+    BotToUser
+}
