+++ src/Moneo.Models/Chat/User.cs
@@ -1,3 +1,3 @@
-namespace Moneo.Bot;
+namespace Moneo.Models.Chat;
 
 public record User(Guid Id, string FirstName, string? LastName, long ConversationId);
