+++ src/Moneo.Models/Chat/UserMessage.cs
@@ -1,3 +1,3 @@
-namespace Moneo.Bot;
+namespace Moneo.Models.Chat;
 
 public record UserMessage(long ConversationId, string Text, string UserFirstName, string? UserLastName = null);
