+++ src/Moneo.Models/TaskManagement/MoneoTaskManagerDto.cs
@@ -1,0 +1,3 @@
+namespace Moneo.Models.TaskManagement;
+
+public record MoneoTaskManagerDto(MoneoTaskState Task, long ChatId);
