+++ src/Moneo.Models/TaskManagement/MoneoTaskResult.cs
@@ -1,0 +1,5 @@
+namespace Moneo.Models.TaskManagement;
+
+public record MoneoTaskResult(bool IsSuccessful, string? ErrorMessage = null);
+
+public record MoneoTaskResult<T>(bool IsSuccessful, T Result, string? ErrorMessage = null);
